requirements.txt
https://github.com/New-dev0/Telethon/archive/patch-8.zip
python-decouple
pyrogram
heroku3